Compress Filtering Job Chain
Convert from chain of IJobParallelFor to a single IJob that creates the resulting sorted lists.
Trade HashMaps for Lists: the cost of allocating/clearing the hashmap ended up more than computing IndexOf() on Lists, for small lists.  The complexity of making a hashmap to convert to a list because NativeQueue has a leak and we don't have IParallelFilterJob from ECS is basically not worth it.  Fully utilizing the multiple cores for burst jobs this small/short was also very spotty.