Fix convectionTempMultiplier
This value is used elsewhere in determining, among other things, the background temperature for the exposed vice unexposed portions of skin.  It can be used similarly to ptd.exposed or skinExposedAreaFrac. Since VehicleOcclusion returns the skin exposed area taking occlusion into account, this value should be 1: the temperature of the air at the exposed area shouldn't be scaled.  However, it needs to be 0 if the entire part is occluded, because of the design choice that skinExposedAreaFrac=1 (instead of 0) in that situation, and so the background temperature of the air at the part needs to interpolate to the atmosphere temperature, not some convection shock temperature.